refactor(grouped-list): дедупликация расчётов виртуального списка
Поведение не менялось — только устранение повторов, накопившихся вокруг якоря
скролла:

- верх контейнера в координатах документа считался четырьмя копиями одного
  выражения (getVirtualItemScrollTop, scrollToItem, watcher захвата якоря,
  watcher resetKey) — теперь одна функция getContainerDocumentTop;
- поиск строки по ключу элемента дублировался в getVirtualItemScrollTop и
  scrollToItem — вынесен в findVirtualRowIndexByItemKey;
- чистые функции без реактивных зависимостей (getColumnCount, getComparableKey,
  sortKeysAuto, sortKeysOrdered) переехали в utils.ts по правилу «Pure
  Functions Outside Closure» из AGENTS.md.

Механика якоря (sync-захват на скролле, восстановление по activeVirtualColumns
с доводкой позиции по кадрам) не тронута сознательно: перенос этой логики в
композабл в прошлых итерациях ломал обратный переход список → сетка.

Идентичность поведения проверена замером до и после (какая карточка вверху
вьюпорта, 3 прогона): /bestiary — круг сетка → список → сетка byte-в-byte
(761/2057), ресайз 1400 → 1180 → 920 → 1400 и /feats совпали по всем ячейкам.
Прогнаны pnpm lint:workflow, stylelint:workflow, type-check — чисто.
